+++ src/i18n/i18n.ts
@@ -3,16 +3,14 @@ import LanguageDetector from "i18next-browser-languagedetector";
 import { initReactI18next } from "react-i18next";
 
 import de from "./locales/de/translation.json";
-import dev from "./locales/dev/translation.json";
 import en from "./locales/en/translation.json";
 import es from "./locales/es/translation.json";
 import fr from "./locales/fr/translation.json";
 
-const languages = ["en", "dev", "es", "fr", "de"];
+const languages = ["en", "es", "fr", "de"];
 
 const resources = {
 	en: { translation: en },
-	dev: { translation: dev },
 	es: { translation: es },
 	fr: { translation: fr },
 	de: { translation: de },